Continuing geometry restructure (#2065)
* I moved calculate_distances to a new folder, packet_transfer. Dependencies were renamed.

* I moved calculate_distances to a new folder, transport. Dependencies were renamed.

* I added myself to the .mailmap

* Moved r_packet_transport from montecarlo_numba to transport. Changed dependencies. Corrected a typo in a comment ("varibale" --> "variable").

* Moved frame_transportations to transport from montecarlo_numba. Renamed dependencies.
