+++ tardis/montecarlo/montecarlo_numba/single_packet_loop.py
@@ -3,7 +3,7 @@ from numba import njit
 from tardis.montecarlo.montecarlo_numba.r_packet import (
     PacketStatus,
 )
-from tardis.montecarlo.montecarlo_numba.r_packet_transport import (
+from tardis.transport.r_packet_transport import (
     move_r_packet,
     move_packet_across_shell_boundary,
 )
@@ -12,7 +12,7 @@ from tardis.montecarlo.montecarlo_numba.continuum.r_packet_transport_continuum i
 )
 
 
-from tardis.montecarlo.montecarlo_numba.frame_transformations import (
+from tardis.transport.frame_transformations import (
     get_inverse_doppler_factor,
     get_doppler_factor,
 )
