+++ tardis/transport/r_packet_transport.py
@@ -12,7 +12,7 @@ from tardis.montecarlo.montecarlo_numba.estimators import (
     update_line_estimators,
     set_estimators,
 )
-from tardis.montecarlo.montecarlo_numba.frame_transformations import (
+from tardis.transport.frame_transformations import (
     get_doppler_factor,
 )
 from tardis.montecarlo.montecarlo_numba.numba_config import (
@@ -71,7 +71,7 @@ def trace_packet(r_packet, numba_model, numba_plasma, estimators):
     )
     comov_nu = r_packet.nu * doppler_factor
 
-    cur_line_id = start_line_id  # initializing varibale for Numba
+    cur_line_id = start_line_id  # initializing variable for Numba
     # - do not remove
     last_line_id = len(numba_plasma.line_list_nu) - 1
